+++ Bullet.h
@@ -7,13 +7,13 @@ class Bullet : public GameObject
 {
 public:
 	Bullet();
-	Bullet(const int _x, const int _y);
+	Bullet(const uint8_t _x, const uint8_t _y);
 	~Bullet();
 
 	void Draw() override;
-	void Update(unsigned long deltaTime) override;
+	void Update(uint8_t deltaTime) override;
 
-	void Init(const int _x, const int _y);
+	void Init(const uint8_t _x, const uint8_t _y);
 
 	bool active;
 };
